requirements.txt
playwright==1.40.0
tkinter-tooltip==2.0.0
cryptography==41.0.7
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pandas==2.1.4
openpyxl==3.1.2 